+++ build.gradle
@@ -45,7 +45,10 @@ dependencies {
     bundle group: 'com.google.guava', name:'guava', version: '19.0'
     bundle group: 'net.engio', name: 'mbassador', version: '1.3.0'
     bundle group: 'com.google.code.gson', name: 'gson', 'version': '2.5'
+
     bundle group: 'com.dmdirc', name: 'com.dmdirc.events.eventbus', version: '+', changing: true
+    bundle group: 'com.dmdirc', name: 'com.dmdirc.util.io.yaml', version: '+', changing: true
+
     bundle group: 'com.dmdirc', name: 'api', version: '+', changing: true
     bundle group: 'com.dmdirc', name: 'util', version: '+', changing: true
     bundle group: 'com.dmdirc', name: 'parser-common', version: '+', changing: true
