+++ bundles/com.dmdirc.util.io.yaml/src/main/java/com/dmdirc/util/io/yaml/BaseYamlStore.java
@@ -15,8 +15,10 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.dmdirc.util;
+package com.dmdirc.util.io.yaml;
 
+import com.esotericsoftware.yamlbeans.YamlReader;
+import com.esotericsoftware.yamlbeans.YamlWriter;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
@@ -28,14 +30,10 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.esotericsoftware.yamlbeans.YamlReader;
-import com.esotericsoftware.yamlbeans.YamlWriter;
-
-import static com.dmdirc.util.YamlReaderUtils.asList;
+import static com.dmdirc.util.io.yaml.YamlReaderUtils.asList;
 import static java.util.stream.Collectors.toList;
 
 /**
@@ -43,17 +41,20 @@ import static java.util.stream.Collectors.toList;
  */
 public abstract class BaseYamlStore<T> {
 
-    /** The charset to use when reading and writing files. */
+    /**
+     * The charset to use when reading and writing files.
+     */
     private static final String CHARSET = "UTF-8";
 
-    /** Logger to use. */
+    /**
+     * Logger to use.
+     */
     private static final Logger LOG = LoggerFactory.getLogger(BaseYamlStore.class);
 
     /**
      * Reads a list of items from the YAML file at the specified path.
-     *
-     * <p>Each item is converted from an object into some form of model by invoking the
-     * {@link #convertFromYaml} method.
+     * <p>
+     * Each item is converted from an object into some form of model by invoking the {@link #convertFromYaml} method.
      *
      * @param path The path of the file to read.
      * @return A list of successfully converted model objects.
@@ -62,7 +63,7 @@ public abstract class BaseYamlStore<T> {
         final List<T> results = new ArrayList<>();
         if (Files.exists(path)) {
             try (final InputStream stream = Files.newInputStream(path);
-                    final InputStreamReader reader = new InputStreamReader(stream, CHARSET)) {
+                 final InputStreamReader reader = new InputStreamReader(stream, CHARSET)) {
                 final YamlReader yamlReader = new YamlReader(reader);
                 results.addAll(asList(yamlReader.read(), this::convertFromYaml));
                 yamlReader.close();
@@ -75,21 +76,20 @@ public abstract class BaseYamlStore<T> {
 
     /**
      * Writes a collection of items to a YAML file at the specified path.
-     *
-     * <p>Each item is converted into a "plain" object by invoking the {@link #convertToYaml}
-     * method.
+     * <p>
+     * Each item is converted into a "plain" object by invoking the {@link #convertToYaml} method.
      *
      * @param path The path of the file to write.
      * @param items A collection of items to write.
      */
     protected void write(final Path path, final Collection<T> items) {
         try (final OutputStream stream = Files.newOutputStream(path);
-                final OutputStreamWriter writer = new OutputStreamWriter(stream, CHARSET)) {
+             final OutputStreamWriter writer = new OutputStreamWriter(stream, CHARSET)) {
             final YamlWriter yamlWriter = new YamlWriter(writer);
             yamlWriter.write(items.parallelStream().map(this::convertToYaml).collect(toList()));
             yamlWriter.close();
         } catch (IOException ex) {
-            LOG.error(LogUtils.USER_ERROR, "Unable to write to {}", path, ex);
+            LOG.warn("Unable to write to {}", path, ex);
         }
     }
