+++ bundles/com.dmdirc.util.io.yaml/src/main/java/com/dmdirc/util/io/yaml/YamlReaderUtils.java
@@ -15,18 +15,14 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.dmdirc.util;
-
+package com.dmdirc.util.io.yaml;
 
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.function.Function;
 import java.util.stream.Collectors;
-
 import javax.annotation.Nullable;
-
-import static com.google.common.base.Preconditions.checkNotNull;
 
 /**
  * Provides utility methods for classes that read from YAML files.
@@ -43,7 +39,6 @@ public final class YamlReaderUtils {
      * suppression to be reduced just to the single line.
      *
      * @param map The wildcard map to be cast.
-     *
      * @return A usable object-to-object map.
      */
     @SuppressWarnings("unchecked")
@@ -57,7 +52,6 @@ public final class YamlReaderUtils {
      * suppression to be reduced just to the single line.
      *
      * @param list The wildcard list to be cast.
-     *
      * @return A usable object list.
      */
     @SuppressWarnings("unchecked")
@@ -69,55 +63,47 @@ public final class YamlReaderUtils {
      * Checks that the specified object is a map, and casts it.
      *
      * @param object The object to be cast.
-     *
      * @return The given object cast as a map.
-     *
      * @throws IllegalArgumentException If the specified object is not a map.
      */
-    public static Map<Object, Object> asMap(@Nullable final Object object)
-            throws IllegalArgumentException {
+    public static Map<Object, Object> asMap(@Nullable final Object object) throws IllegalArgumentException {
         if (object instanceof Map) {
             return uncheckedCast((Map<?, ?>) object);
         }
 
-        throw new IllegalArgumentException("Unexpected element. Found "
-                + simpleName(object) + ", expected Map");
+        throw new IllegalArgumentException("Unexpected element. Found " + simpleName(object) + ", expected Map");
     }
 
     /**
      * Checks that the specified object is a list, and casts it.
      *
      * @param object The object to be cast.
-     *
      * @return The given object cast as a list.
-     *
      * @throws IllegalArgumentException If the specified object is not a list.
      */
-    public static List<Object> asList(@Nullable final Object object)
-            throws IllegalArgumentException {
+    public static List<Object> asList(@Nullable final Object object) throws IllegalArgumentException {
         if (object instanceof List) {
             return uncheckedCast((List<?>) object);
         }
 
-        throw new IllegalArgumentException("Unexpected element. Found "
-                + simpleName(object) + ", expected List");
+        throw new IllegalArgumentException("Unexpected element. Found " + simpleName(object) + ", expected List");
     }
 
     /**
      * Checks that the specified object is a list, and casts it.
      *
      * @param object The object to be cast.
-     * @param convertor The function to use to convert each list member to the desired type.
-     *
+     * @param converter The function to use to convert each list member to the desired type.
      * @return The given object cast as a list.
-     *
      * @throws IllegalArgumentException If the specified object is not a list.
      */
-    public static <T> List<T> asList(@Nullable final Object object,
-            final Function<Object, Optional<T>> convertor) throws IllegalArgumentException {
+    public static <T> List<T> asList(@Nullable final Object object, final Function<Object, Optional<T>> converter)
+            throws IllegalArgumentException {
         return asList(object).parallelStream()
-                .map(convertor).filter(Optional::isPresent)
-                .map(Optional::get).collect(Collectors.toList());
+                .map(converter)
+                .filter(Optional::isPresent)
+                .map(Optional::get)
+                .collect(Collectors.toList());
     }
 
     /**
@@ -125,16 +111,11 @@ public final class YamlReaderUtils {
      *
      * @param map The map to retrieve the entry from.
      * @param key The key of the entry to retrieve.
-     *
      * @return The string representation of the value of the key.
-     *
      * @throws IllegalArgumentException If the specified key is not present, or is empty.
      */
     public static String requiredString(final Map<Object, Object> map, final String key)
             throws IllegalArgumentException {
-        checkNotNull(map);
-        checkNotNull(key);
-
         if (!map.containsKey(key)) {
             throw new IllegalArgumentException("Required key not present: " + key);
         }
@@ -152,14 +133,10 @@ public final class YamlReaderUtils {
      *
      * @param map The map to retrieve the entry from.
      * @param key The key of the entry to retrieve.
-     *
      * @return The string value of the entry with the given key, or {@code null} if not present.
      */
     @Nullable
     public static String optionalString(final Map<Object, Object> map, final String key) {
-        checkNotNull(map);
-        checkNotNull(key);
-
         final Object value = map.get(key);
         return value == null ? null : value.toString();
     }
@@ -168,7 +145,6 @@ public final class YamlReaderUtils {
      * Gets the simple name of the given object's class.
      *
      * @param object The object to get the class name of
-     *
      * @return The simple class name of the object, or the string {@code "null"}.
      */
     private static String simpleName(@Nullable final Object object) {
